commandes admin + velo + declinaison utilisable fonctionnel

+++ controllers/admin_velo.py
@@ -18,22 +18,31 @@ admin_velo = Blueprint('admin_velo', __name__,
 @admin_velo.route('/admin/velo/show')
 def show_velo():
     mycursor = get_db().cursor()
+
     sql = ''' 
-     SELECT nom_velo AS nom,
-         id_velo AS id_velo,
-         libelle_type_velo AS libelle_type_velo,
-         velo.id_type_velo AS type_velo_id,
-         libelle_taille AS libelle_taille,
-         velo.id_taille AS taille_id,
-         prix_velo AS prix,
-         stock AS stock,
-         photo AS image
-     FROM velo
-         INNER JOIN type_velo ON velo.id_type_velo = type_velo.id_type_velo
-         INNER JOIN taille ON velo.id_taille = taille.id_taille
-          '''
+     SELECT V.nom_velo AS nom,
+            V.id_velo AS id_velo,
+            V.id_type_velo AS type_velo_id,
+            V.prix_velo AS prix,
+            V.utilisable AS utilisable,
+            COALESCE(V.image, 'pas d\\'image') AS image,
+            T.libelle_type_velo AS libelle_type_velo,
+            SUM(D.stock) AS stock,
+            MIN(D.stock) AS min_stock,
+            COUNT(D.id_declinaison_velo) AS nb_declinaisons
+            
+     FROM velo AS V
+         
+     JOIN type_velo T ON V.id_type_velo = T.id_type_velo
+     LEFT JOIN declinaison_velo D ON V.id_velo = D.id_velo
+         
+     GROUP BY V.id_velo, V.nom_velo, V.utilisable, V.id_type_velo, V.prix_velo, V.image, T.libelle_type_velo
+         
+     ORDER BY V.nom_velo ASC
+  '''
     mycursor.execute(sql)
     velos = mycursor.fetchall()
+
 
     return render_template('admin/velo/show_velo.html', velos=velos)
 
@@ -65,7 +74,6 @@ def valid_add_velo():
     type_velo_id = request.form.get('type_velo_id', '')
     prix = request.form.get('prix', '')
     description = request.form.get('description', '')
-    stock = request.form.get('stock', '')
     image = request.files.get('image', '')
 
 
@@ -76,11 +84,11 @@ def valid_add_velo():
         print("erreur")
         filename=None
 
-    tuple_add = (None, nom,  prix, description, filename, stock, 1, type_velo_id)
+    tuple_add = (None, nom,  prix, description, filename,type_velo_id)
 
     sql = '''
-    INSERT INTO velo (id_velo, nom_velo, prix_velo, description, photo, stock, id_taille, id_type_velo) VALUES
-          (%s, %s, %s, %s, %s, %s, %s, %s)'''
+    INSERT INTO velo (id_velo, nom_velo, prix_velo, description, image, id_type_velo) VALUES
+          (%s, %s, %s, %s, %s, %s)'''
 
     mycursor.execute(sql, tuple_add)
     get_db().commit()
@@ -97,36 +105,62 @@ def valid_add_velo():
 def delete_velo():
     id_velo=request.args.get('id_velo')
     mycursor = get_db().cursor()
-    sql = '''SELECT
-                 0 AS nb_declinaison,
-                '' AS image
-             FROM velo WHERE id_velo = %s'''
+
+
+    sql = '''SELECT COUNT(D.id_declinaison_velo) AS nb_declinaison
+          FROM velo 
+          INNER JOIN declinaison_velo D ON velo.id_velo = D.id_velo
+         WHERE velo.id_velo = %s AND D.utilisable = TRUE
+          '''
     mycursor.execute(sql, id_velo)
     nb_declinaison = mycursor.fetchone()
+
+
     if nb_declinaison and nb_declinaison['nb_declinaison'] > 0:
         message= u'il y a des declinaisons dans cet velo : vous ne pouvez pas le supprimer'
         flash(message, 'alert-warning')
     else:
-        sql = ''' SELECT photo AS image FROM velo WHERE id_velo = %s'''
+        # verifier l'utilisation du velo
+        sql = '''
+            SELECT COUNT(*) AS nbr_utilisations
+                
+            FROM velo 
+            JOIN note        N ON velo.id_velo = N.id_velo
+            JOIN commentaire C ON velo.id_velo = C.id_velo
+            JOIN liste_envie L ON velo.id_velo = L.id_velo
+            JOIN historique  H ON velo.id_velo = H.id_velo
+            LEFT JOIN declinaison_velo D ON velo.id_velo = D.id_velo
+                
+                
+            WHERE velo.id_velo = %s
+        '''
         mycursor.execute(sql, id_velo)
-        velo = mycursor.fetchone()
-        image = velo['image']
+        nbr_utilisations = mycursor.fetchone()['nbr_utilisations']
 
-        sql = '''DELETE FROM ligne_panier WHERE id_velo = %s'''
-        mycursor.execute(sql, id_velo)
+        if int(nbr_utilisations) >= 1:
+            sql = '''
+                UPDATE velo SET utilisable = FALSE
+                    WHERE id_velo = %s
+            '''
+            mycursor.execute(sql, id_velo)
+            get_db().commit()
 
-        sql = '''DELETE FROM ligne_commande WHERE id_velo = %s'''
-        mycursor.execute(sql, id_velo)
+        else:
+            sql = ''' SELECT image AS image FROM velo WHERE id_velo = %s'''
+            mycursor.execute(sql, id_velo)
+            velo = mycursor.fetchone()
+            image = velo['image']
 
-        sql = ''' DELETE FROM velo WHERE id_velo = %s'''
-        mycursor.execute(sql, id_velo)
-        get_db().commit()
-        if image != None:
-            os.remove('static/images/' + image)
 
-        print("un velo supprimé, id :", id_velo)
-        message = u'un velo supprimé, id : ' + id_velo
-        flash(message, 'alert-success')
+            sql = ''' DELETE FROM velo WHERE id_velo = %s'''
+            mycursor.execute(sql, id_velo)
+            get_db().commit()
+            if image != None:
+                os.remove('static/images/' + image)
+
+            print("un velo supprimé, id :", id_velo)
+            message = u'un velo supprimé, id : ' + id_velo
+            flash(message, 'alert-success')
 
     return redirect('/admin/velo/show')
 
@@ -137,10 +171,12 @@ def delete_velo_cascade():
     id_type_velo = request.args.get('id_type_velo')
     mycursor = get_db().cursor()
 
-    sql = '''SELECT 0  AS nb_declinaison, \
-                    '' AS image
-             FROM velo \
-             WHERE id_velo = %s'''
+    sql = '''
+          SELECT COUNT(D.id_declinaison_velo) AS nb_declinaison
+          FROM velo 
+          INNER JOIN declinaison_velo D ON velo.id_velo = D.id_velo
+         WHERE id_velo = %s
+     '''
     mycursor.execute(sql, id_velo)
     nb_declinaison = mycursor.fetchone()
 
@@ -148,22 +184,12 @@ def delete_velo_cascade():
         message = u'il y a des declinaisons dans cet velo : vous ne pouvez pas le supprimer'
         flash(message, 'alert-warning')
     else:
-        sql = ''' SELECT photo AS image \
+        sql = ''' SELECT image AS image \
                   FROM velo \
                   WHERE id_velo = %s'''
         mycursor.execute(sql, id_velo)
         velo = mycursor.fetchone()
         image = velo['image']
-
-        sql = '''DELETE \
-                 FROM ligne_panier \
-                 WHERE id_velo = %s'''
-        mycursor.execute(sql, id_velo)
-
-        sql = '''DELETE \
-                 FROM ligne_commande \
-                 WHERE id_velo = %s'''
-        mycursor.execute(sql, id_velo)
 
         sql = ''' DELETE \
                   FROM velo \
@@ -185,15 +211,26 @@ def edit_velo():
     id_velo=request.args.get('id_velo')
     mycursor = get_db().cursor()
     sql = '''
-        SELECT  nom_velo AS nom,
-         id_velo AS id_velo,
-         id_type_velo AS type_velo_id,
-         prix_velo AS prix,
-         stock AS stock,
-         photo AS image,
-            description AS description
-        FROM velo
-        WHERE id_velo = %s
+        SELECT V.nom_velo                   AS nom,
+               V.description                AS description,
+               V.id_velo                    AS id_velo,
+               V.id_type_velo               AS type_velo_id,
+               V.prix_velo                  AS prix,
+               V.image                      AS image,
+               T.libelle_type_velo          AS libelle_type_velo,
+               SUM(D.stock)                 AS stock,
+               MIN(D.stock)                 AS min_stock,
+               COUNT(D.id_declinaison_velo) AS nb_declinaisons
+        
+        FROM velo AS V
+            
+        JOIN type_velo T ON V.id_type_velo = T.id_type_velo
+        LEFT JOIN declinaison_velo D ON V.id_velo = D.id_velo
+            
+        WHERE V.id_velo = %s
+            
+        GROUP BY V.id_velo, V.nom_velo, V.description, V.id_type_velo, V.prix_velo, V.image, T.libelle_type_velo
+        
     '''
 
     mycursor.execute(sql, id_velo)
@@ -207,16 +244,32 @@ def edit_velo():
     mycursor.execute(sql)
     types_velo = mycursor.fetchall()
 
-    # sql = '''
-    # requête admin_velo_6
-    # '''
-    # mycursor.execute(sql, id_velo)
-    # declinaisons_velo = mycursor.fetchall()
+    sql = '''
+        SELECT D.id_velo             AS velo_id,
+               D.id_declinaison_velo AS id_declinaison_velo,
+               D.id_taille           AS id_taille,
+               D.id_couleur          AS id_couleur,
+               D.stock               AS stock,
+               D.utilisable          AS utilisable,
+               C.libelle             AS libelle_couleur,
+               T.libelle             AS libelle_taille
+            
+        FROM declinaison_velo as D
+            
+        JOIN couleur C ON D.id_couleur = C.id_couleur
+        JOIN taille T ON D.id_taille = T.id_taille
+            
+        WHERE D.id_velo = %s
+            
+        ORDER BY D.utilisable DESC
+    '''
+    mycursor.execute(sql, id_velo)
+    declinaisons_velo = mycursor.fetchall()
 
     return render_template('admin/velo/edit_velo.html'
                            ,velo=velo
                            ,types_velo=types_velo
-                         #  ,declinaisons_velo=declinaisons_velo
+                           ,declinaisons_velo=declinaisons_velo
                            )
 
 
@@ -231,7 +284,7 @@ def valid_edit_velo():
     description = request.form.get('description')
     stock = request.form.get('stock')
 
-    sql = ''' SELECT photo AS image FROM velo WHERE id_velo = %s'''
+    sql = ''' SELECT image AS image FROM velo WHERE id_velo = %s'''
     mycursor.execute(sql, id_velo)
     image_nom = mycursor.fetchone()
     image_nom = image_nom['image']
@@ -246,37 +299,9 @@ def valid_edit_velo():
             image_nom = filename
 
     print(type_velo_id)
-    sql = '''  UPDATE velo SET nom_velo = %s, photo = %s, prix_velo = %s, velo.id_type_velo = %s, description = %s , stock = %s WHERE id_velo = %s '''
-    mycursor.execute(sql, (nom, image_nom, prix, type_velo_id, description, stock, id_velo))
+    sql = '''  UPDATE velo SET nom_velo = %s, image = %s, prix_velo = %s, velo.id_type_velo = %s, description = %s WHERE id_velo = %s '''
+    mycursor.execute(sql, (nom, image_nom, prix, type_velo_id, description, id_velo))
 
-
-    sql = '''
-        SELECT COALESCE(COUNT(ligne_panier.id_velo), 0) AS nbr_paniers, COALESCE(SUM(quantite), 0) AS nbr_velos_panier
-        FROM ligne_panier
-        WHERE id_velo = %s
-    '''
-    stock = int(stock)
-    mycursor.execute(sql, id_velo)
-    result = mycursor.fetchone()
-    nbr_velos_panier = int(result['nbr_velos_panier'])
-    nbr_paniers = int(result['nbr_paniers'])
-
-    if stock < nbr_velos_panier:
-        left = nbr_paniers - stock
-
-        sql = '''
-            SELECT quantite, id_utilisateur, id_velo
-            FROM ligne_panier
-            WHERE id_velo = %s
-        '''
-        mycursor.execute(sql, id_velo)
-        paniers = mycursor.fetchmany(nbr_paniers)
-
-        for panier in paniers:
-            sub = math.ceil((panier['quantite'] / nbr_velos_panier) * left)
-            sql = ''' UPDATE ligne_panier SET quantite = quantite - %s WHERE id_velo = %s AND id_utilisateur = %s '''
-            mycursor.execute(sql, (sub, id_velo, panier['id_utilisateur']))
-            left -= sub
 
 
     get_db().commit()
